DRF1/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-extensions==3.2.3
ipython==8.29.0